+++ changes.md
@@ -5,6 +5,8 @@ Dart Web Toolkit Showcase Changes
 
 * All code has been changed use IoC and MVP.
 * Start using prettify instead of highlite.
+* Changed prettify.css to hide border in code prettify
+* Create examples: Html, Hyperlink
 
 **March 10, 2013